fix : V4 타임라인 테이블 마이그레이션 오류 수정 및 타입 최적화

+++ backend/src/main/java/com/ssafy/s309/domain/prediction/entity/GlucosePrediction.java
@@ -32,7 +32,7 @@ public class GlucosePrediction {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Integer id;
+  private Long id; // Integer에서 Long로 변경
 
   @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "user_id", nullable = false)
